@@ -5,45 +5,41 @@
   "main": "app/src/main.tsx",
   "dependencies": {
     "copy-to-clipboard": "^3.3.1",
-    "core-js": "^3.6.5",
+    "core-js": "^3.12.1",
     "normalize.css": "^8.0.1",
     "query-string": "Heapy/query-string",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "regenerator-runtime": "^0.13.7",
-    "typestyle": "^2.1.0"
+    "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
     "@types/history": "^4.7.8",
-    "@types/node": "^14.11.2",
-    "@types/react": "^16.9.49",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-router": "^5.1.8",
-    "@types/react-router-dom": "^5.1.5",
-    "autoprefixer": "^10.0.0",
-    "copy-webpack-plugin": "^6.1.1",
-    "css-loader": "^4.3.0",
-    "file-loader": "^6.1.0",
-    "html-webpack-plugin": "^4.5.0",
-    "less": "^3.12.2",
-    "less-loader": "^7.0.1",
-    "postcss": "^8.0.9",
-    "postcss-loader": "^4.0.2",
-    "style-loader": "^1.2.1",
-    "ts-loader": "^8.0.4",
-    "typescript": "^4.0.3",
-    "url-loader": "^4.1.0",
-    "webpack": "^4.44.2",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0"
+    "@types/node": "^15.0.2",
+    "@types/react": "^17.0.5",
+    "@types/react-dom": "^17.0.3",
+    "@types/react-router": "^5.1.14",
+    "@types/react-router-dom": "^5.1.7",
+    "autoprefixer": "^10.2.5",
+    "copy-webpack-plugin": "^8.1.1",
+    "css-loader": "^5.2.4",
+    "html-webpack-plugin": "^5.3.1",
+    "less": "^4.1.1",
+    "less-loader": "^8.1.1",
+    "postcss": "^8.2.14",
+    "postcss-loader": "^5.2.0",
+    "style-loader": "^2.0.0",
+    "ts-loader": "^9.1.2",
+    "typescript": "^4.2.4",
+    "webpack": "^5.37.0",
+    "webpack-cli": "^4.7.0",
+    "webpack-dev-server": "^3.11.2"
   },
   "scripts": {
     "prestart": "yarn install",
     "prebuild": "yarn install",
-    "start": "webpack-dev-server --mode development --env.NODE_ENV development --env.SOURCE_MAP eval-source-map",
-    "build": "webpack --mode production -p --env.NODE_ENV production --env.SOURCE_MAP source-map"
+    "start": "webpack serve --mode development --env NODE_ENV=development --env SOURCE_MAP=eval-source-map",
+    "build": "webpack --mode production --env NODE_ENV=production --env SOURCE_MAP=source-map"
   },
   "repository": {
     "type": "git",
